refactor: EventService의 이벤트 관리 중앙화
