+++ src/main/java/christmas/domain/service/EventService.java
@@ -4,7 +4,6 @@ import christmas.domain.customer.Customer;
 import christmas.domain.customer.VisitDate;
 import christmas.domain.eventplanner.Event;
 import christmas.domain.eventplanner.EventBadge;
-import christmas.domain.eventplanner.EventPlanner;
 import christmas.domain.restaurant.Gift;
 import christmas.domain.restaurant.Orders;
 import christmas.dto.EventPreviewDto;
@@ -15,20 +14,23 @@ import java.util.Map;
 
 public class EventService {
 
-    private final EventPlanner eventPlanner;
-
-    public EventService() {
-        this.eventPlanner = new EventPlanner();
+    public List<Event> findEventsByCustomer(Customer customer) {
+        return Event.from(customer);
     }
 
-    public Gift getGiftMenu(Customer customer) {
-        return eventPlanner.getGift(customer);
+    public Gift gatherEligibleEventGifts(Customer customer) {
+        List<Event> events = findEventsByCustomer(customer);
+
+        return events.stream()
+                .filter(Event::hasGift)
+                .map(Event::getGift)
+                .reduce(Gift.empty(), Gift::merge);
     }
 
     public Map<Event, Integer> getBenefitDetails(Customer customer) {
         Map<Event, Integer> benefitDetails = new HashMap<>();
 
-        List<Event> events = eventPlanner.findEventsByCustomer(customer);
+        List<Event> events = findEventsByCustomer(customer);
         for (Event event : events) {
             benefitDetails.put(event, event.calculateBenefit(customer));
         }
@@ -42,7 +44,7 @@ public class EventService {
     }
 
     private int getBenefitAmount(Customer customer) {
-        List<Event> events = eventPlanner.findEventsByCustomer(customer);
+        List<Event> events = findEventsByCustomer(customer);
 
         return events.stream()
                 .mapToInt(event -> event.calculateBenefit(customer))
@@ -51,9 +53,9 @@ public class EventService {
 
     public EventPreviewDto getEventPreviewDto(VisitDate visitDate, Orders orders) {
         Customer customer = Customer.reserveVisit(visitDate, orders);
-        Gift gift = getGiftMenu(customer);
+        Gift gift = gatherEligibleEventGifts(customer);
         Map<Event, Integer> benefitDetails = getBenefitDetails(customer);
-        List<Event> events = eventPlanner.findEventsByCustomer(customer);
+        List<Event> events = findEventsByCustomer(customer);
         int discountAmount = getDiscountAmount(events, customer);
         EventBadge badge = getBadgeForBenefitAmount(customer);
